+++ CHANGELOG.md
@@ -10,8 +10,11 @@
     * Rename `handleAppSwitchReturnURL()` to `handleReturnURL()`
     * Rename `canHandleAppSwitchReturnURL()` to `canHandleReturnURL()`
     * Remove `unregisterAppSwitchHandler()`
+  * Replace `BTVenmoDriver.authorizeAccount` methods with `BTVenmoDriver.tokenizeVenmoAccount`
 * Fix memory leak in `BTPayPalDriver`
 * Add `offerPayLater` to `BTPayPalRequest`
+* Add `environment` to `BTConfiguration`
+* Add `BTVenmoRequest`
 
 ## 5.0.0-beta2 (2021-01-20)
 * Add SPM support for `BraintreeDataCollector` and `BraintreeThreeDSecure`